requirements.txt
dify_plugin>=0.2.0,<0.3.0
pillow==11.3.0